+++ src/app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -19,7 +19,7 @@ class AuthenticatedSessionController extends Controller
     public function create(): Response
     {
         return Inertia::render('Auth/Login', [
-            'canResetPassword' => Route::has('password.request'),
+            'canResetPassword' => true,
             'status' => session('status'),
             'isFirstUser' => !\App\Models\User::exists(),
         ]);